pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ongarde"
version = "1.0.0"
description = "OnGarde — Runtime content security layer for self-hosted AI agent platforms"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"

dependencies = [
    # Core proxy
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.27.0",

    # Audit backend
    "aiosqlite>=0.20.0",

    # Utilities
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "python-ulid>=2.2.0",  # ULID generation for X-OnGarde-Scan-ID and audit scan_id

    # Logging
    "structlog>=24.1.0",
]

[project.optional-dependencies]
full = [
    "presidio-analyzer>=2.2.361",
    "spacy>=3.8.0",
    # After install: python -m spacy download en_core_web_sm
    "google-re2>=1.1",
    "bcrypt>=4.1.0",
    "watchfiles>=0.21.0",
    "supabase>=2.4.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "anyio[trio]>=4.0.0",
    "httpx>=0.27.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
ongarde = "app.run:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

# Uvicorn hardened defaults (architecture.md §9.3):
# uvicorn app.main:app \
#   --host 127.0.0.1 \
#   --port 4242 \
#   --limit-concurrency 100 \
#   --backlog 50 \
#   --timeout-keep-alive 5 \
#   --timeout-graceful-shutdown 30
